Split the OpenGL wrapping layer out, part 1.
First just moving files to a new location and updating their include
guards to reflect the new location, but *not* updating the namespace
just yet. Creating a new MagnumGL library out of it and linking
everything that depends on GL stuff to this library instead of just
Magnum. At the moment it's basically everything except Math (of course),
SceneGraph and Shapes. It's awful, but it will get better over time -- I
need to do it in small steps to keep this in a reviewable condition.

In particular, even the image wrapper classes are now part of the GL
wrapping layer (because they use GL definitions), so also all of Trade
and transitively all importer plugins depend on it. This is the most
important problem to solve in order to be able to load plugins using
Vulkan.

For compatibility the original headers are still present in the root
namespace, but marked as deprecated and not installed in the
non-deprecated build.